backend/vendor_microservice/requirements.txt
Flask==3.1.2
python-dotenv==1.1.1
boto3==1.40.5

# Testing
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-env==1.2.0
